// workspaces/github-discussions/plugins/search-backend-module-github-discussions/src/types.ts
/*
 * Copyright 2024 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { IndexableDocument } from '@backstage/plugin-search-common';

export interface GithubDiscussionsDocument extends IndexableDocument {
  author: string;
  category: string;
  labels: {
    name: string;
    color: string;
  }[];
  comments: {
    author: string;
    bodyText: string;
    replies: {
      author: string;
      bodyText: string;
    }[];
  }[];
}